Add gh-pages url to cors allowlist
Currently experimenting with moving the frontend over to Github Pages.  
This commit adds the github url to our cors policy, so we can use the backend api during testing. 

If we decide to switch over to gh-pages, then we'll point the `antalamanc.com` url at github and can remove `icssc-projects.github.io/AntAlmanac`.